Changed over to using net instead of http
The submission server tests weren't working
using the http abstraction as it's just for
non-persistent url requests.

net is the underlying socket implementation
of the http package, both of which are built
in to node and can be used separately.
By using net instead of http the submission
test server managed to work as I could respond
raw data vs http requests.

Also removed all the semi-colons.